Use supervisor instead of nodemon for development server
nodemon was not detecting changes because of a misconfiguration.
However, it is also unstable in use on our development vm.
The vm system time gets out of sync with the host system time.
nodemon seems to use the modified date of files to detect changes
and when the times get out of sync, nodemon continuously restarts
the app.