Show quarterly ticks on x-axis for monthly data over long timeframe.
When we have monthly data going back more than 15 months, use quarterly ticks.

I've over-written the "month" config in graph.js. We don't seem to be using
this anywhere in the live app, or testing it, so I think it's OK to overwrite.
But I'd be grateful if someone with more Spotlight experience could check.