Add currency symbols to graph.
Committing on a clean branch, this supersedes the previous branch.

Part of the G-Cloud story: https://www.pivotaltracker.com/s/projects/911874/stories/62377934

Provide a configurable "currency" option in Stagecraft. If set, use currency symbols in y-axis labels and callouts.

Allow both for currency prefixes (like "£10") and possible future suffixes (like "10 CNY").

If the currency option is set, use one fewer decimal place in the numbers shown in callouts, so that the text does not overflow as a result of having an extra character. This doesn't account for the possible future use of suffixes, but I suggest we look at that when we get there.