Add rightHandGracePixels to hover.js
These few pixels increase the size of the hover area for the middle
section of a graph with linelabels to the right.

Before: as soon as you move from the final date value of the graph
into the linelabel, the hover disappears.

Now: the first few leftmost columns of pixels of the linelabel are
assumed to be hovering on the graph still.