Remove code for propositional header to fix wiggle
The code for the propositional navigation and global header has been
moved into the govuk_template. Removed it from the project. Also the new
correct place to put the propositional name is in the propositionHeader
block not insideHeader.

Removed the `menu` skip link from the navigation as there is no
navigation to skip to.