Add GroupedTimeshiftCollection to timeshift data
For some graphs we want to be able to show a series timeshifted from
the other series. Add a collection which will increse the duration
required to cover all the series. Then slice the return values so the
correct parts of data are selected for each series.

Allow the grouped_timeseries module to automatically switch to this
collection if it detects configuration asking for a timeshifted series.