Refactored single_stat.js to fix graph errors.
When 'change:selected' we should only work with the model that's emitted.

If we're not on a line (eg a part of the graph with no data the matrix emits an array) so we now have some logic to test for an array.
Previously it was doing [].get('something') which was throwing exceptions. We now test for that logic and pick the first model from that array.