Return an array of an array from tickValues on conversion graph.
In commit cf5dabaa5edff06db0e8c6f321def0b7e4be209a we started using
.apply to allow us to pass multiple values to D3 functions. In this case
we want to treat the results of tickValues as a single argument, not as
multiple values to be passed. Hence, return an array of an array.

Discussed this with Tom Booth and we think this is the least-worst fix.