VisitorsRealtimeCollection -> ListCollection
We are going to need basically the same functionality for gathering and
displaying data for the ListView, so rather than writing something new I
thought I would generify this class.

All functionality as expected by visitors-realtime should have remained
the same, I have adapted and not removed any of the tests.