Timeseries, sort by _timestamp if no _start_at
When porting VisitorsRealtimeCollection to ListCollection I left in
data.reverse(). This was used to rearrange the data returned from
Backdrop into the correct order. Timeseries already tried to rearrange
the order of data based on _start_at, but realtime data about visitors
only has _timestamp. As ListCollection is now more generic, the job of
ordering should be the responsibility of Timeseries (probably should
have been before).

I have changed Timeseries to prioritise ordering on _start_at but if it
doesn't exist then fall back to _timestamp. Tests as proof!