Add an optional regex to extract label
We have some text we want to strip off from the labels we get back from
Backdrop. I have added a label-regex parameter to the module which gets
used in ListView. If it exists to will be converted to a RegExp object
and exec'd on the label, the first group in the regex will be extracted.