Add better stagecraft config and column styling
I've limited the column layout for modules down to just .list at the
moment as it is the only module which .cols2 is avaliable.

I have added the stagecraft config but the trending config is still
pointing at the wrong data, this will be changed when the data is
available.

I added the list module to the Stagecraft linter such that it would
still do its great job and lint all the things.