Improve naming of startMatcher/endMatcher.
The use of 'startMatcher' and 'endMatcher' is technical debt from when
the completion collections were only used to calculate the percentage
of users ending a transaction, compared to the number of users starting
the transaction.

Now we want to use them to calculate more generic percentages - e.g.
the percentage of transactions that are digital. So it is confusing to
use 'start' and 'end' when what we really mean is 'denominator' and
'numerator'.

Rename accordingly.

Part of this story: https://www.pivotaltracker.com/s/projects/911874/stories/63460908