Rename .tickValues to .tickValueList for clarity.
Currently we set a property called .tickValues on our D3 scale. This
is confusing, because there is an official D3 property called .tickValues
that is used on axes. There is no such property on D3 scales.

Rename this to .tickValueList for clarity: we are storing a list of
tick values, so this name is more descriptive and less likely to be
confused with the D3 axis property.