Load steps from axes.y for journey modules.
For consistency with other modules with `axes` property in configuration when rendering tabular data.

Default `key` property of journey step to `uniqueEvents` to avoid repetitive input in module config.
The reason uniqueEvents is the property for all journeys is due to the collect query that we make to backdrop.
When it collapses events it will collate the numbers as uniqueEvents.