Allow user satisfaction to take extra classes.
Now we can set .cols2 and have half-width user satisfaction modules.

Note that this makes the "Report a problem" link reposition itself oddly,
higher up the page. This is because we should only ever have two of these
modules in place, not one.

This demonstrates the problem: http://jsfiddle.net/4Xhwe/ and the fix:
http://jsfiddle.net/4Xhwe/1/

We could fix this by adding "clear: both" to .report-a-problem-toggle, but
since we should never actually have only one .cols2 module in place, I'm not
sure if I should edit a global style just to fix this temporary issue.

Also add user satisfaction data to SORN dashboard.