Fix bug in delta.js.
If the delta was very small, but non-zero, previously delta.js showed
"0.00%" with an upwards or downwards arrow, rather than "no data".

Fix this to ensure that delta.js looks at the displayed value, not the
absolute underlying value, when working out what to display.

Also, we didn't have any tests for displaying .no-change. Add this, and
add tests to ensure that it's displayed correctly.

There's another bug in formatPercentage - it shouldn't ever return "-0.00%",
with a minus symbol. I've ticketed this separately.