Fix to sparklines with tables.
Updated the css to position 'info element' absolute to get round aligment issues.

We add this before the table as more of a 'summary' for the graph.

This allows us to add a table after the graph and not have any issues with the 'summary' moving to the bottom of the table.