Date ranges for x-axis in tables
In order to handle ranges we need to be able to pull out a composite key
which is then converted into a string for display in the format fn.

I've had to drop the non-array form of getTableRow as if the x-axis can
be an array it causes the conversion of args -> array to fail.

I've also had to move the 'flattening' of rows across multiple series.
Previously this was happening in the view, but again the requirement for
an array of values would cause the view logic to fail due to looking for
arrays as values. I've moved the 'flattening' into the logic in
MatrixCollection, it will only return the x-axis value for the 0th
collection and the other vals will be flattened into a row.