Add floatHeaders method to table view
Change the table to render on the client-side, so that we can float
the headers properly to give a nice scroll experience. It would be nicer
to only render this small part of the table client-side, but for now
we have to render the whole thing.

Add a floatHeaders method that sets the width of the th and the first
row of td elements, and then sets a class on the table.