Add styles to floated-header tables
Once we've set the column width correctly in JavaScript, we can make
use the class that was then applied to the table to ensure that it
has a maximum height and scrolls for longer content.

Because tables are rendered with JavaScript on the client-side, we need
a fallback visualisation in old browsers. Set raw styles so that we get
a screenshot of the whole table.

I know :(