Add Sass to support having two-column modules
At the top of the page only at this time. This is because it's really
difficult to know whether you're dealing with a section on the left
side or right side without introducing additional config.

Browser support for nth-of-type is the same as first-of-type. IE9+
is the big one for us, but I don't consider that a problem.