Remove NTA tabular numbers from header
NTA tabular numbers causes an issue in Chrome on Windows that we
first encountered in 358e541c5e4d1ab873fa0bd3d7ea8dd1aa76549b

https://code.google.com/p/chromium/issues/detail?id=344347

This is particularly visible in the page header. We should never
need to display fixed width numbers in the header anyway.

Thanks to @ianhopkinson for the bug report.