Use compass to compile image-path in the toolkit
I know more Ruby isn't the direction we want to be going on, but
the govuk_frontend_toolkit contains image-path attributes that are
currently not being compiled by Sass.

This commit adds Compass to the build process for Sass, and sets
some Compass configuration correctly so that images can be
included in the CSS and cachebusted.