Share module rendering and merge view options
We are going to want to share the rendering of modules beyond just the
dashboard controller. The new tab module is built around the idea of is
containing another level of module definitions.

We want the options we pass through to render ending up on the view such
that we can pass through useful properties like, which element to render
into from elsewhere.