Use existing attributes on query change.
This is important for tabbed modules. It ensures that if a start or
end date is set by the initial query, we will continue to use it
on subsequent queries made by other tabs.

The only exception is the availability tabs, which switch period,
and where we therefore want to reset the start date according
to the period.

NB: All this will be superseded by the "tabbable modules" story,
due to be checked in shortly. It is an interim fix.