Fix bug in live-service-usage graph rendering.
We were checking for the existence of the collection before
rendering the template HTML for live-service-usage graphs.

Unfortunately, this got us into an occasional race condition
between templateContext and initialize, so sometimes the
graphs weren't rendering.

The template check isn't really useful anyway, because we
don't have error handling in the collection yet, so the
graphs will probably break before this point if there is
no data in the collection.

So: remove this check from the template, and remove
templateContext from the view.