Fix lists not become 100% width when on mobile
There was an issue on production where the list module were not changing
to fill the complete width of page when the width was reduced.

It turns out that we had column code knocking around in a few places in
our scss codebase. I have collapsed them down to one set of styling
columns shared between the homepage and modules.