Fix bug in completion rate modules: pass filter params.
There was a bug in the student finance dashboard: filter parameters
had been specified for the completion rate modules, but were not
being passed through to the query. Fix this and update tests.

Interestingly, it looks as though the completion module does have
an option to set query parameters directly on the module, though
this wasn't being used. However, I suggest just using this quick
fix for now, and following up by reviewing the use of query
parameters across the whole of Spotlight - it would be good to
rationalise the way we handle them.