Links should be blue.
The underline might be enough in other circumstances, but here it looks like we're just underlining the heading.

I was prompted today - Matt mentioned that Lindsey from Student Finance
hadn't noticed the module pages existed at all. I think this will go some
way to fixing that problem, and is better than adding a 'more details' link
to every module.

Next, blue clashes hard with the blue-green, so we also lose the blue-green.