Show previous value if current value unavailable in KPIs.
In the KPI module, if the current value of the data is unavailable,
currently we only display "no data" in the headline, and nothing
at all in the section where the delta value usually goes.

It would be nicer to display the previous value of the data in
the section where the delta value usually goes, so if the data
is out of date, at least the user has some information about
its previous value.

Amend KPI module to do this.