Add Transaction Explorer modules back in again
They had been removed for a production release and are now ready to
drop back in again.

The modules for Carer's Allowance and LPA will be merged separately
since the data is not yet ready.

Don't include data for pay-legalisation-drop-off or for
pay-foreign-marriage-certificates since these are not in TxEx. Also
not for deposit-foreign-marriage since it has been archived.