Refactor journey graph styles
Move the media query into the meat of the Sass, because it doesn't
actually need to be a media query. It's also not specific enough
to be used where it previously was.

Hide the tick line on the journey graph. It was never in any designs
so I'm not sure why it's there.

Add a tiny bit more space between the arrow and the journey graph.