Fix bug in journey styles: callout too wide
The callout under the journey graph is too wide, resulting in the edges
being cropped.

This commit makes the callout 1% smaller on both the left and the right,
and then moves the arrow 1% to the left to compensate (so it's still
centered horizontally).

Also add a bit more space to the top and the bottom, because it's
all quite bunched up.