Improve colour management in stylesheets
Use Sass variables to reference colours. Ideally we'd get
rid of the mixing I've added and just use standard greys,
but for now it's still nicer to use proper variables and
not just magic hex values.

Remove unused $primary-colour-text variable.

There shouldn't be any noticeable difference in the browser
as a result of this commit.