Amend graph.js to handle datasets in which all data is null.
Set a default max value in calcYScale to handle the case in which
all data in a dataset is null.

Previously, without a default max value, graphs in which all data
was null would throw an error when trying to calculate this scale.

Now the graph will show 'no data' on all its points, and the y-axis
will be set to a small default value.

This is needed for the new bar_chart_with_number module, which has
explicit designs for graphs in which all data is null. Plus it
seems generally beneficial that graphs should cope with all-null
datasets.