This fixes rendering tabs when no js is available on the browser.
Basically we have to force the rendering of tabs onto the client because the sub views of them may require svg.

To ensure alphagov/screenshot-as-a-service gets a rendered module we change the ready event to a loaded and fire it when all sub modules are loaded.

Also had to move the tab tests to client only due to the ```requiresSvg``` stuff.