Add bar_chart_with_number module.
Add module, collection, view and template files for our
new module.

We expect 'axis-period' to be specified on this module,
since it is always a time series (by default, we set it
to 'week). We use the axis-period property to define
the start_date set in the collection, and the labels shown
on the chart.