Default values in journey-series collection
When backdrop does not have any hits at all for one of the steps in a journey collection it returns no data point, which means the graph bar is rendered with a "no data" label.

If we have data for some of the steps, then assume that the value for the missing step should be zero, rather than "no data".