Allow controllers to bootstrap their collections with data
We often fetch the data for a collection on the server, so there is little value in having the client fetch it again, when it can be bootstrapped into the module.

Allow controllers to define a `collectionData` method, which if it returns a non-empty array, will prevent the collection from fetching again on the client, and will render immediately.