Content dashboards shouldnt appear in service list
At the moment the list of services includes all dashboards, which means
the numbers on the homepage don't add up and we are showing non-service
dashboards in the list of service dashboards.

I've renamed services.json to dashboards.json as this more truely
conveys what it is. This change knocks on into the collection.

I've made filterDashboards able to accept an array so that we can store
a list of dashboard-types which are 'services' and reference that
constant in any code that wants to filter all services.