Add content dashboards to the homepage
I switched the styling on the 1/3 + 2/3 columns to use the correct grid
styles, so that the widths were uniform down the whole homepage.

I had to do a bit a nasty in homepage.js in order to filter out the
GOV.UK content dashboard, which meant hardcoding a plus one into the
template. Not a massive fan but the other option I thought of would have
been to filter the list in the template which I liked even less.

I removed the hardcoded size from the transactions explorer tree-map
image as it seemed a little weird and I couldn't find a reason for its
existence.

The add3 style did not include the extra margins width it removed and
would never will 100% of the width. I have rectified this and pulled out
the widths to vars to be shared between col3 and add3.

The 'a strong' styling was added to make the two links on the homepage
match the style of the designs.

Conflicts:
	styles/common/columns.scss