Adjusted dates in kpi tests.
The dates now reflect real data that would be returned from backdrop.

A year would be Jan - Jan so we should reflect that.

A week is actually 8 days so we should reflect that.

This should help anyone who's wondering why we subtract a day off a week and a month off a year.