Ensure that department/agency names aren't repeated in HV dashboards.
For some transactions, the department name is duplicated in the agency
column in the TxEx spreadsheet.

We can't change this for the moment because it's used in Transactions
Explorer. So let's add a fix in the Spotlight import script not to
add the agency name if it matches the department name.

Pivotal story: https://www.pivotaltracker.com/s/projects/911874/stories/71010818

In addition, this commit contains some changes made by Cliff:

* "Citizens" become "Individuals" throughout
* Minor text changes from TxEx spreadsheet