Remove ymin config
this wasn't really actually doing anything useful. It was only used in sparklines, which were only used for real-time usage graphs. These always ended up with a min y value of 0, in the early hours of the morning, and it seems that we would always want these to be scaled from zero anyway.