Add published flag to all files with correct val
We can no longer be so coarsed grained as to say the all of experimental
counts as 'unpublished' as deposit foreign marriage needs to stay in
place but be unpublished.

This model is nicer than experimental/ as it means we can leave
dashboards in place and just toggle whether they are visible on the
homepage.