Move strapline to start of <title> tag.
This is so that our "Dashboard" strapline is more visible in Google
search results - at the moment it's often hidden at the end of
a long service name. This results in users clicking through
to the dashboards when really they're looking for the service
itself.

Story: https://www.pivotaltracker.com/s/projects/911874/stories/71237550