Add meta description for transaction pages.
This is for SEO purposes, in an effort to reduce the number of users who
click through to our transaction pages by mistake after entering generic
transaction-related search terms.

The meta description tries to make it clearer that this is a performance
dashboard with statistics - not the service itself.

Background here: https://www.pivotaltracker.com/s/projects/911874/stories/71237550