The 24 hour filtering was the wrong way round
The problem reared its head when we stopped gathering realtime data for
around 8 hours before restarting. It was chopping the most recent data
off the front of the list of points.

I've switched the tests to use the correct collection so it exhibits the
same sorting and structure as the view will receive. I have also
reordered the data to reflect the order it comes back from backdrop.