Show six bars on 'bar chart with number' module.
This is so that we can show six quarters of data by default on the
high-volume bar charts.

Almost all of the high-volume services actually have data going
back six quarters (to Q4 2012) and the more we can show, the
more meaningful the graphs become.

Change font-size on the x-axis labels to 12px: this matches the
font size used for x-axis labels in journey graphs.