Extend grouped_timeseries to calculate percentages.
We want the option of displaying percentages on line graphs with
multiple series. Extend the grouped_timeseries module to allow
this.

Unlike the stacked graphs, where D3 does the percentage calculations
for us, we need to calculate the percentages ourselves for line
graphs.

Also store the original values of the data - we'll need this to
display on line labels and in callouts.