Switch the category of visitors module to website
This module had previously been abusing the dataType field returned by
backdrop. The dataType for all of the records should be identical as the
data-set should have only one dataType. This commit will switch this
graph to using the new field website which should allow correct
switching on the data.