Format time-taken graph values as minutes/seconds, not raw seconds.
The GA data is returned to us in seconds, but we want to show it
as minutes and seconds.

Parse it into minutes in the collection, then pass display options
through to the number and graph views.

It might be cleaner to keep it in seconds, but I wasn't sure how
to get the y-axis to return a neat set of axis ticks formatted
in minutes, without changing the underlying data.

Needs tests.