Allow users to specify optional steps in journey graphs.
At the moment, if there are optional steps in our journey graphs, it's
confusing since they tend to have less traffic than steps that are
later in the journey, but there is no visual indication that these steps
are optional.

So, highlight optional steps in grey.

No tests, since it's firebreak!