Add various exit-rate-related stats to the CA stub data.
"exits" and "exitRate" we get back from the GA API.

Ideally we would use "exitRate" on our graphs, however I think this is
an average rate that has been aggregated across the pagePathLevel1
(i.e at the section level) and so it doesn't look right.

Instead I've calculated exitsOverUsers and added it via a script
- this is the raw number of exits recorded, over the raw number of
users, for each section level.