Calculate the exit rate.
GA does give us an exitRate statistic, but it doesn't work well when
aggregated across sections.

For sections, we need to look at the total number of exits anywhere in
that section, over the total number of users in that section. (I think...
Need to get this peer reviewed. It might be over the total number of
page views.)

Anyway, this is a little script to calculate that and add it to the
stub data.

If we decide to deploy this, we may need to add this calculation
to the GA collector for this statistic.