Move data request properties into data-source
We want to contain all of the properties related to how a module
requests data from backdrop into one object. This changes the schema for
modules to fold up this data into one object.

The schema for period looks a bit weird with its allOf wrapping. It is
there because if you have an enum it will ignore whether it is required
so inorder to coax our jsonschema lib to validate correctly I had to
push separate these rules into an allOf.