Fix module rendering tests
Use inspection for particular DOM elements of interest where possible, instead of matching enormous html strings.

Remove `fallback` property from view config since it isnt actually used anywhere.

Remove `offerPng` property as it is effectively a duplicate of `fallbackUrl` - it is true if and only if there is a fallbackUrl defined.