Pin Carer's completion rate module to 19 May
Instead of displaying data going back 9 weeks, just display
back to the 19th of May because that's when we started collecting
data with the `thank-you` label.

This change should be reverted in a few weeks once there is enough data
in place to make a decent-looking graph. A chore has been created for that:

https://www.pivotaltracker.com/story/show/72963084