Remove multi-stats module from codebase.
We no longer use this module anywhere and we have no plans to use
it anywhere. So we have agreed to remove it entirely.

Some of the sub-views used by multi-stat items were actually used
elsewhere in the codebase, so I have kept these but moved them
out of multi-stat-specific directories. The affected views are
sparkline, headline and delta views, and these are used by the
user satisfaction and realtime modules.

Note that spec.realtime.js now throws a D3 console error in the
Jasmine tests. I'm not sure why this is: the graphs are rendering
fine in the browser.