Make tables scrollable on touch devices.
Add a .touch-table class to tables on touch devices.

When this class is set, add a 1.5em left margin, and reduce
the width of the table accordingly.

I think calc() is the best way to do this, since we can't use
border-box with a margin, and we can't use padding on the table
since it has a border. The only other way I can see it do it
is a containing div and that starts to get messy.

Update tests to check that the class is set.