Add department filtering to services page
This commit adds a select box to the services page to filter
services based on their department.

It generates an array of departments using a dashboard collection
which it uses to populate the select box. We create slugs based on
the department's abbreviation.

This works in browsers with and without JavaScript by adding a
submit button to the form.