Update the module schema to use a data-source
The allOf in the period property on the query-params is needed to get
around the fact that an enum cannot be required=false in our jsonschema
lib. By defining it this way we can force this to work.

Schemas that explicitly required a period have been dropped and in the
case of availability removed altogether as it only added that field.