Migrate config to use data-source parameters
I have written a script that migrates the config
(tools/migrate-config-to-data-source.js), it works by picking out any
parameters which we formally used by the collections (both loose and
inside of query-params).

The script will also try and migrate some of the defaulted or derived
values into the config to be more explicit and decouple the data-source
from other parts of spotlight.