Retrofit data sources into modules
We can remove a lot of the queryParam functions as the data-source
parameters in modules now contain the required parameters. This means
they are already known about in the collections rather than needing to
be passed through. Maps are maintain when modules need to collect
specific data.

The period should now be exposed through the collection rather than the
modules delving around in their options.

The realtime module should always cut off at 24 hours, so rather than
storing a phoney period and duration hardcode it.