Tables should use strings
During the performance analysis of spotlight we realised that ever
request spent a lot of CPU and memory rendering tables. It was shown to
be the use of the 'DOM' that caused these performance issues. Switching
the rendering to assemble strings and then parse them in one go has
reduced the mean response time, on local loadtest, by a factor of 5.

The tests have been modified to consider the output of render rather
than messing around with the intermediary functions.