Fix Collection tests
queryId is not a thing anymore so we can just fallback to the Backbone
fetch fn. This also means the tests got nuked. I looked this param up
and it was created in Limelight for some suprious reason involving
storing query parameters based on a query id. We definitely don't do
that anymore.

There was a bunch of duplication in the url tests where it treated
multiple arbitary params differently to params like filter_by. I don't
know why they should be considered differently and they both follow the
same path in the code now, so I killed the tests.