Fix JourneyCallout tests
This should have been using the correct collection to test against, this
was shown up because the JourneySeriesCollection stores the dataRange it
will request.

A set of axes needs to be defined for the collection to be able to
flatten the data.

We need to shim getMoment so we can fixed what now is so the tests don't
go crazy.