Fix JourneySeriesCollection tests
I had to remove the prototype valueAttr so that it could be varied.
Previously if passed in through the config it would not get set because
the base Collection will only set it if there is not a valueAttr
property already set on this.

The dateRange for the collection is now only calculated on init so the
moment munging needs to move to before the use of the constructor.