Sortable tables https://www.pivotaltracker.com/n/projects/911874/stories/67381806
This requires the matrix free collections.
The client template modules now require the collection in order for the sorting.
All modules have been enabled to use the table sort.
Lots of tests.
Table render on the backend has been modified very slightly to take a collection as an argument instead of just assuming it's the this.collection (as we use this.tableCollection to get round sorting data that then gets updated in the graphs!)
Supports secondary sorting for dates (_timestamp and _start_at)
Includes minor styling.