Use a time scale in graph.js for duration data
Extend graph.js to use a time scale for data in duration format.
Set the domain of the scale using D3's .nice() function, not
by supplying an explicit domain and tick value list.

Supply the appropriate time interval to .nice() by looking
at the maximum value of the data.