Fix table rendering where X axis is label
If we are making a groupped query and do not have time as the X axis,
then we do not want to try and collapse the data.

This is the case in tables where the first column (x axis) is a label
against the sum of a particular field.

The line and stacked graph tests had to change as the data was expecting
to be collapsed but it needed a date as the x axis to make this happen.