Improve tab styles on small screens
Tabs weren't really styled at all on small screens.

This commit rejigs some of the styles so that they are more mobile
first, and makes them a little bit tidier on small screens (they're
nowhere near perfect, but improved).

In some cases I haven't styled this page mobile first (eg background
and border colours). This is to improve the page for browsers that
don't support media queries (IE 8) because it doesn't complicate
our code that much.