Stop cheapseats running on every Travis run
We are going to move cheapseats to run async on every successful release
tag. At the moment it is slowing down our pipeline from code -> preview
a lot. If something falls through the cracks post removing cheapseats
from the pipeline we can revisit its presence.

We no longer need to split up the build into 3 parts as we are dropping
cheapseats.