Highlight the last point with data on a line when hovering a line label
DRY out the event handling code a little, so the label link hover handler and `onHover` use the same method to select collection items.

Test against a spy on `Collection#selectItem` instead of the side-effects of that method to avoid testing too much of the collection internals.

Have to allow the label positions to be set to set-up the tests, so need to pass a bunch more stub config in.