Use header tag rather than section for unpublished banner
Fixes this bug:
https://www.pivotaltracker.com/s/projects/537859/stories/75018342

This occured because our `:first-of-type` selector was matching the
banner, not the first KPI module. CSS3 doesn't have a way to select
the first element with a particular class, so it's easier just to
change the element type for the banner.

In addition, header is arguably a more appropriate tag for the
unpublished banner.