Group data before calculating "other" and "total" fields
When applying group mapping, we need to do it before we calculate the "other" field because datasets might get grouped into defined groups.

Only include data groups without a defined line in total figures if there is also a defined line for "other", to ensure the numbers displayed to a user always add up to the total.

Support regular expressions for group mappings.