Log backend requests
By including it in the requirements for the bootstrap to be called we
can be certain that it has been loaded before we start running code.

This protect against potential race conditions when requiring client
logger.

Also added a no-op logger to test scripts via helper.

Needed to add new global logger to jshint globals/predefs.