Implements a fallback on the stagecraft api model fetch method
When this is called it will first attempt to get config from stagecraft
before falling back to a request to the local app.

This is tested by faking sync to record the url used at each stage in
the fallback. There is also a unit test for the fallback toggling
itself.

This still requires pp-deployment changes to set up the per environment
urls